Moved stats collection to per thread context.
Prior to this change, the scalability of memcachetest was inhibited by
a big lock around the stats collection.  With this change, the lock is
no more and stats collection is done per thread, consolidating and
summarizing at the end of the run.

The percentile distribution was removed, as it needs to be aware of the
changes.  Worry not.  It will be back.